+++ templates/InviteTranslations/index.php
@@ -20,7 +20,7 @@
                     <tr>
                         <td style="padding: 5px" align="center"><?= $this->Number->format($inviteTranslation->sortOrder) ?></td>
                         <td style="padding: 5px"><?= h($inviteTranslation->language->name) ?></td>
-                        <td style="padding: 5px"><?= $this->Html->link(__(h($inviteTranslation->subject)), ['action' => 'view', $inviteTranslation->id]) ?></td>
+                        <td style="padding: 5px"><?= $this->Html->link(__('[DH Course Registry] ' .h($inviteTranslation->subject)), ['action' => 'view', $inviteTranslation->id]) ?></td>
                         <td style="padding: 5px" align="center"><?= ($inviteTranslation->active) ? 'Yes' : 'No' ?></td>
                         <td class="actions" style="padding: 5px" align="center"><?= $this->Html->link(__('Edit'), ['action' => 'edit', $inviteTranslation->id]) ?></td>
                     </tr>